+++ starsky/starsky.foundation.accountmanagement/Middleware/UseBasicAuthenticationExtensions.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.AspNetCore.Builder;
+
+namespace starsky.foundation.accountmanagement.Middleware
+{
+    public static class UseBasicAuthenticationExtensions
+    {
+
+        public static IApplicationBuilder UseBasicAuthentication(this IApplicationBuilder builder)
+        {
+            return builder.UseMiddleware<BasicAuthenticationMiddleware>();
+        }
+    }
+}
