+++ starsky/starsky.foundation.accountmanagement/Models/Account/LoginViewModel.cs
@@ -1,0 +1,18 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace starsky.foundation.accountmanagement.Models.Account
+{
+    public class LoginViewModel
+    {
+        [Required]
+        [EmailAddress]
+        public string Email { get; set; }
+
+        [Required]
+        [DataType(DataType.Password)]
+        public string Password { get; set; }
+
+        [Display(Name = "Remember me?")] 
+        public bool RememberMe { get; set; } = true;
+    }
+}
