+++ starsky/starsky/Attributes/PermissionAttribute.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
-using starskycore.Services;
+using starsky.foundation.accountmanagement.Services;
 
 namespace starsky.Attributes
 {
