+++ starsky/starsky/Views/Account/Register.cshtml
@@ -1,4 +1,4 @@
-﻿@model starskycore.ViewModels.Account.RegisterViewModel
+﻿@model starsky.foundation.accountmanagement.Models.Account.RegisterViewModel
 @{
     ViewData["Title"] = "Register";
 }
