+++ starsky/starskyadmincli/starskyadmincli.csproj
@@ -11,9 +11,9 @@
     </PropertyGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\starsky.foundation.accountmanagement\starsky.foundation.accountmanagement.csproj" />
       <ProjectReference Include="..\starsky.foundation.database\starsky.foundation.database.csproj" />
       <ProjectReference Include="..\starsky.foundation.platform\starsky.foundation.platform.csproj" />
-      <ProjectReference Include="..\starskycore\starskycore.csproj" />
     </ItemGroup>
 
 </Project>
