--- starsky/starskycore/Middleware/MiddlewareExtensions.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.AspNetCore.Builder;
-
-namespace starskycore.Middleware
-{
-    public static class MiddlewareExtensions
-    {
-
-        public static IApplicationBuilder UseBasicAuthentication(this IApplicationBuilder builder)
-        {
-            return builder.UseMiddleware<BasicAuthenticationMiddleware>();
-        }
-    }
-}
