--- starsky/starskycore/ViewModels/Account/LoginViewModel.cs
@@ -1,18 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace starskycore.ViewModels.Account
-{
-    public class LoginViewModel
-    {
-        [Required]
-        [EmailAddress]
-        public string Email { get; set; }
-
-        [Required]
-        [DataType(DataType.Password)]
-        public string Password { get; set; }
-
-        [Display(Name = "Remember me?")] 
-        public bool RememberMe { get; set; } = true;
-    }
-}
