+++ starsky/starskytest/Helpers/ConfigCliAppsStartupHelperTest.cs
@@ -28,22 +28,10 @@ namespace starskytest.Helpers
 			name = cliHelper.ReadMeta().ToString();
 			Assert.AreEqual(true,name.Contains(nameof(ReadMeta)));
 			
-			// name = cliHelper.ImportService().ToString();
-			// Assert.AreEqual(true,name.Contains(nameof(Import)));
-
-			name = cliHelper.UserManager().ToString();
-			Assert.AreEqual(true,name.Contains(nameof(UserManager)));		
 			
 			name = cliHelper.ThumbnailCleaner().ToString();
 			Assert.AreEqual(true,name.Contains(nameof(ThumbnailCleaner)));		
 		}
 
-		// [TestMethod]
-		// [ExpectedException(typeof(Microsoft.EntityFrameworkCore.Storage.RetryLimitExceededException))]
-		// public void ConfigCliAppsStartupHelperTestImportService_MysqlCrash()
-		// {
-		// 	Environment.SetEnvironmentVariable("app__DatabaseType","mysql");
-		// 	new ConfigCliAppsStartupHelper();
-		// }
 	}
 }
