+++ starsky/starskytest/Interfaces/IUserManagerTest.cs
@@ -1,5 +1,6 @@
 using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using starsky.foundation.accountmanagement.Interfaces;
 using starsky.foundation.database.Models.Account;
 using starskycore.Interfaces;
