+++ starsky/starskytest/Middleware/BasicAuthenticationSignInManagerTest.cs
@@ -5,6 +5,9 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using starsky.foundation.accountmanagement.Interfaces;
+using starsky.foundation.accountmanagement.Middleware;
+using starsky.foundation.accountmanagement.Services;
 using starsky.foundation.database.Data;
 using starsky.foundation.platform.Models;
 using starskycore.Interfaces;
